// web/src/shared/api/generated/model/studentUnitSummaryResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface StudentUnitSummaryResult {
  id: string;
  name: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  lessonCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  servableCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  masteredCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  seenCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  masteryPercent: number | string;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
     */
  bestExamScorePercent: number | string | null;
}